Extract workspace chrome and boot UI
